Move MessageContext to dispatch::context module
Move the MessageContext to the dispatch module, in order to reduce the
dispatcher module's complexity.
